Fix circular import of bw version check
